Restructuring the app.
I change the structure of components. Some components are just copied from thanapoom-dechalamai/sbacprofile_v2.

There's now a use of useContext. As for now, I don't know that this will make me less insane or me banging my head on the desk later down the line.

There's also the use of i18next for multi-language app. It's being tested but the functionality is to be expected.