UI Redesign
The background color of the app is now changed.

The rolodex card now has the white color with a small shadow.

In rolodexes, I decided to use the background of the image to display the major color instead of the card color. The contacts are now removed from the card, the user can still see the contact information within the modal.