:white_check_mark: Huge issue fixed!
There's a persisting bug since the sbacprofile_v2.
The issue is that if the user is already logged in. When they close the app, and came back to it. The userInfo state is empty, causing the app to break, the user has to log out in the order for that userInfo to be fetched again.
It's fixed by making the app fetches the userInfo every time the app is initailized, and do the re-render after all is done to make the dashboard link available.

Other changes:
- Changed the primary color from purple to iOS default blue.
- Sidebar now has the background based on the user major.
- Small changes that was supposed to made with the commit that comes before this one.