User club dashboard/page
There's now a page that displays relevant information about the user's club.

I went to all the component that I've made and removed the unnecessary API calls. Contexts now has it's predefined fetch functions for the consumer to call upon to update the state. The fetch will only be called when nothing is in the state.

The API fetch will not be called in the parent component, and will be called in the component that need access to it instead. There's a check if the info already exists, of course.

There's also a bug where if the user creates or updates a student or teacher, if there's no image, the image the info still gets uploaded. I've added a simple return block, just to stop the info from being updated.

Small changes:
- Many table components will no longer render the same thing conditionally. Meaning that the table header will be rendered, the condition only comes in the table content.
- An announcement rolodex card has been made for the home page to use that component.
- Loading component will no longer use react-cssfx-loading, and will use Font Awesome icon instead.